backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
motor==3.3.2
pydantic==2.6.1
email-validator==2.1.0
resend==2.4.0
reportlab==4.1.0
requests==2.31.0